vaccine-checker/requirements.txt
requests==2.24.0
python-dotenv==0.19.0
urllib3==1.25.9